Fixing overlapping function name for Solaris
This function overshadows the one from libc on Solaris, and guess what,
is hence never called.  Since the original one has an entirely different
contract, it results in unconditional segfaults later on, so always use
the internal copy as it's small enough not to care.

Gentoo patch 0070_all_solaris-basename.patch